pyproject.toml
[project]
name = "x-algorithm-enhancements"
version = "0.1.0"
description = "Enhancements to xAI's open-sourced recommendation algorithm: KV-cache optimization, reward modeling, and multimodal retrieval"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core JAX stack (Phoenix dependencies)
    "jax>=0.4.30,<0.5.0",
    "jaxlib>=0.4.30,<0.5.0",
    "dm-haiku>=0.0.12",
    "optax>=0.2.2",
    # Data processing
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    # F1: Multimodal (CLIP)
    "transformers>=4.40.0",
    "torch>=2.2.0,<2.3.0",
    # Visualization & profiling
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-xdist>=3.5.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["enhancements", "phoenix"]

[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = ["phoenix", "home-mixer"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "N806",  # uppercase variable names (N, M, K) — conventional in math/matrix code
    "E402",  # module-level imports not at top — unavoidable in scripts with sys.path setup
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["phoenix/", "home-mixer/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
